fix race condition in factory that allowed to return null
the compare-and-set! call doesn't guarantee it is our logger that has been
changed in the state map. If we don't want to change the logger in the state map
we should call swap! with update and use the result.

As a safe-guard I've put a throw in. Returning null from getLogger caused me to
debug for 4 hours where the issue lies. This will help others to pinpoint us in
case the code messes up again :)